requirements.txt
opencv-python>=4.7.0.72
pyautogui>=0.9.53
pillow>=9.5.0
pyyaml>=6.0
numpy>=1.24.2